Added a download log button
The context menu for text files not originally on the display (i.e user needs to scroll down to view the text file) disappears before the user can click on the menu option. So I added a download log button in the "uploader" that downloads the currently viewed text file.